pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "country_by_country"
version = "0.1.0"
description = "Template"
authors = ["DataForGood"]
license = " MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6.0.1"
camelot-py = "^0.11.0"
opencv-python-headless = "^4.9.0.80"
ghostscript = "^0.7"
pypdf = "^4.0.2"
unstructured = {version="^0.12.6", python =">=3.10,<3.12"}
pdf2image = "^1.17.0"
pillow-heif = "^0.15.0"
unstructured-inference = "^0.7.24"
pytesseract = "^0.3.10"
pikepdf = "^8.13.0"
unstructured-pytesseract = "^0.3.12"
joblib = "^1.3.2"
scikit-learn = "1.2.2"
streamlit = "1.31.1"
llama-parse = "^0.3.9"
python-dotenv = "^1.0.1"
langchain-openai = "^0.1.0"
langchain = "^0.1.13"

[tool.poetry.group.dev.dependencies]
pre-commit = "^2.20.0"
pytest = "^7.2.0"
# ruff = "^0.0.254"
tox = "^4.4.8"
notebook = "^7.1.1"

[tool.ruff]
select = [
    "A",
    "ANN",
    "ARG",
    "B",
    "C",
    "COM",
    "C90",
    "E",
    "ERA",
    "F",
    "I",
    "N",
    "PGH",
    "PL",
    "PLE",
    "PLR",
    "PLW",
    "PT",
    "PTH",
    "Q",
    "RET",
    "RUF",
    "SLF",
    "SIM",
    "TID",
    "TRY",
    "UP",
    "W",
    "YTT",
]
ignore = ["A003", "ANN101", "ANN401", "N805", "N818", "PLR0913", "RET504", "RET505"]

fixable = [
    "A",
    "ANN",
    "ARG",
    "B",
    "C",
    "COM",
    "C90",
    "E",
    "ERA",
    "F",
    "I",
    "PGH",
    "PL",
    "PLE",
    "PLR",
    "PLW",
    "PT",
    "PTH",
    "Q",
    "RET",
    "RUF",
    "SLF",
    "SIM",
    "TID",
    "TRY",
    "UP",
    "W",
    "YTT",
]
unfixable = []

exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "venv",
    "pages",
    "tests"
]

line-length = 96

dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

target-version = "py310"

[tool.ruff.mccabe]
max-complexity = 10
